Move system_status_consumer from controllers to system
