+++ src/system/system_status_consumer.h
